Jo, tim: herorganisatie van de packages + renames
