+++ src/main/java/be/swsb/fiazard/bestelling/Bestelling.java
@@ -2,8 +2,8 @@ package be.swsb.fiazard.bestelling;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static org.apache.commons.lang.StringUtils.isNotBlank;
-import be.swsb.fiazard.event.Aggregate;
-import be.swsb.fiazard.event.AggregateId;
+import be.swsb.fiazard.ddd.Aggregate;
+import be.swsb.fiazard.ddd.AggregateId;
 
 class Bestelling implements Aggregate {
