+++ src/main/java/be/swsb/fiazard/ddd/Aggregate.java
@@ -1,4 +1,4 @@
-package be.swsb.fiazard.event;
+package be.swsb.fiazard.ddd;
 
 public interface Aggregate {
