+++ src/main/java/be/swsb/fiazard/ddd/AggregateIdGenerator.java
@@ -1,4 +1,4 @@
-package be.swsb.fiazard.event;
+package be.swsb.fiazard.ddd;
 
 public class AggregateIdGenerator {
