+++ src/main/java/be/swsb/fiazard/ddd/DomainEvent.java
@@ -1,0 +1,7 @@
+package be.swsb.fiazard.ddd;
+
+
+public interface DomainEvent {
+    int getVersion();
+
+}
