--- src/main/java/be/swsb/fiazard/event/Event.java
@@ -1,7 +1,0 @@
-package be.swsb.fiazard.event;
-
-
-public interface Event {
-    int getVersion();
-
-}
